self.set: Queue for setting values
Reason for this is that a listener updating the same value should
be 'queued' when setting the value, so other listeners will trigger
with the original set - not the new. Or else the previous value would
be sent to listeners, not the new one.
This might only work for 1 "level", only tested for that.