Hello 2.0
Started working on the next iteration of data.
Next version will be strict, but allow for wildcard.
The hope is that with better understanding of what I need in domdom
I can re-structure the library to be leaner, cleaner and meaner.

Wildcards and recursive wildcards

Remove alias

For now. Personally no more need

Started on data

Stuff!

.

.

.

.

Call listeners _after_ setting values. ey

.

.

.

.

.

  .

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.