Hello transformers
Plan is to say bye-bye to pathifier and hello to only-transformers.
The idea is that every transformation is controlled by a transformer.
Main API change from before is that .map will return a new
transformer with the given return type, giving very good
type support!
And it will support transformer branching, making it possible
to split into multiple outputs.